add: teamID in team page & improve card visuals
